Fix Smash ball on app icon and remove icon masks
This fixes the Smash ball logo on the app icon and removes the icon masks (the rounded corners) as they are no longer needed. It also simplifies the filenames of the icons.

For anyone who needs to make changes or export additional icons, my vector recreation is included too (I used Inkscape).

Fingers crossed my changes to Contents.json do not break backward compatibility with older versions of Xcode. And normally the author is supposed to be the bundle ID but it seems to work fine so what the heck, might as well save a few bytes.