Fix app icons again, shrink Smash ball and minify
This tweaks the icon colors, fixes icons that were previously left unfixed due to their masks (I believe these ones are used for the Debian package), shrinks the Smash ball so it fits neatly inside the iOS download circle, minifies the SVG even further, removes unnecessary icon sizes and adds dark and tinted icons for iOS 18 users.

The icon mask uses a radius of 1.25, which seems to be correct for iOS 8 and up (older radii are untested and out of scope). You can add this in Inkscape by selecting the yellow rectangle, then selecting the rectangles and squares tool (R key) and changing the horizontal radius (Rx) and vertical radius (Ry) in the toolbar.

The tinted icon uses the grayscale filter in Inkscape with the following settings:
- Red: 0.24
- Green: 0.84
- Blue: 0.07
- Lightness: 0.16